The problem with deleting
I do not know how to correctly check
if an element is the last in collisions chain
so I cannot mark deleted elements as 0:0.

In this version, due to incorrect check
if element is the last in collisions chain,
the chain of collisions is break and the
element that is in the table is not found.